# feat: optimize search performance with pg_trgm indexes for Japanese and English support