--- a/.gitignore
+++ b/.gitignore
@@ -172,3 +172,4 @@ regenerate.log
 # progress files
 .regenerate-low-item-summaries-progress.json
 .archive/
+.claude/