--- a/package.json
+++ b/package.json
@@ -157,7 +157,7 @@
     "@types/bcryptjs": "^2.4.6",
     "@types/sanitize-html": "^2.16.0",
     "axios": "^1.12.1",
-    "bcryptjs": "^2.4.3",
+    "bcryptjs": "^3.0.2",
     "cheerio": "^1.1.2",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",