--- a/prisma/migrations/20251004194303_add_pg_trgm_indexes_for_japanese_search/migration.sql
+++ b/prisma/migrations/20251004194303_add_pg_trgm_indexes_for_japanese_search/migration.sql
@@ -6,15 +6,17 @@ CREATE EXTENSION IF NOT EXISTS pg_trgm;
 
 -- Create trigram indexes on title and summary columns
 -- These indexes will significantly speed up LIKE/ILIKE queries
--- Note: CONCURRENTLY cannot be used in migration (transaction block)
+-- CONCURRENTLY is used to avoid ACCESS EXCLUSIVE locks in production
 
-CREATE INDEX IF NOT EXISTS idx_article_title_trgm
+CREATE INDEX CONCURRENTLY IF NOT EXISTS idx_article_title_trgm
   ON "Article" USING gin(title gin_trgm_ops);
 
-CREATE INDEX IF NOT EXISTS idx_article_summary_trgm
+CREATE INDEX CONCURRENTLY IF NOT EXISTS idx_article_summary_trgm
   ON "Article" USING gin(summary gin_trgm_ops);
 
 -- Note:
 -- - Existing to_tsvector indexes are not removed (can be used for English-specific advanced search)
 -- - Trigram indexes support all languages including Japanese
 -- - PostgreSQL will automatically use these indexes for ILIKE queries
+-- - IMPORTANT: For production deployment, run with PRISMA_MIGRATION_ENGINE_SKIP_TRANSACTIONS=1
+--   to allow CONCURRENTLY to work outside transaction block (see README.md)