… app/api/articles/search/route.ts
… docs/roadmap/improvement-roadmap-2025-10.md