## Summary

Optimized search performance by adding pg_trgm indexes, achieving 99.9% improvement while maintaining full Japanese and English support.

### Key Changes

- Add pg_trgm trigram indexes for ILIKE search optimization
- Remove broken SQLite FTS5 search API
- Re-enable and enhance E2E search tests
- Update bcryptjs to 3.0.2 for security improvements

### Performance Improvements

| Metric | Before | After | Improvement |
|--------|--------|-------|-------------|
| English search | 60s | 2-67ms | 99.9% |
| Japanese search | 60s | 2-4ms | 99.9% |
| Average search time | 60s | 10.6ms | 99.98% |

### Test Results

- Type check: PASS
- Lint: PASS
- Build: PASS
- Unit tests: 197/197 (100%)
- E2E tests: 12/12 (100%)

### Technical Details

**Database Changes**:
- Added idx_article_title_trgm (GIN index)
- Added idx_article_summary_trgm (GIN index)
- Migration: 20251004194303_add_pg_trgm_indexes_for_japanese_search

**Code Changes**:
- Removed: app/api/articles/search/route.ts (broken FTS5 code)
- Modified: __tests__/e2e/specs/home.spec.ts (re-enable test, add Japanese test)
- Added: scripts/test/measure-search-performance.ts (performance validation)

**Security Updates**:
- bcryptjs: 2.4.3 → 3.0.2 (fully compatible)
- nodemailer: On-hold due to next-auth peer dependency (^6.6.5)

### Language Support

- English: Fully supported, optimized
- Japanese: Fully supported, optimized
- Mixed keywords: Supported (AND search behavior)

## Test plan

- [x] Type check (no errors)
- [x] Lint (no errors)
- [x] Build (success)
- [x] Unit tests (197/197 passed)
- [x] E2E tests (12/12 passed)
- [x] Performance validation (all < 500ms)
- [ ] Staging environment verification
- [ ] Production deployment

## Related Documents

- Investigation: .claude/docs/investigate/investigate_20251004_184945_roadmap_high_priority_tasks.md
- Implementation Plan: .claude/docs/plan/plan_20251004_191817_roadmap_implementation_revised.md
- Implementation Details: .claude/docs/implement/implement_20251004_195410_roadmap_tasks_completed.md
- Test Results: .claude/docs/test/test_20251004_205125_roadmap_implementation.md
- Roadmap: docs/roadmap/improvement-roadmap-2025-10.md

Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- Refactor
  - サーバー側の検索エンドポイントを削除（検索周りのバックエンド構成が変更）。
  - 日本語を含む検索のためのデータベースインデックス最適化で検索体感速度を改善。
- Tests
  - 検索ボックスのE2Eテストを再有効化。
  - 日本語キーワードでの検索動作を検証する新テストを追加。
- Documentation
  - 今後6か月の改善ロードマップを追加（検索最適化、セキュリティ、キャッシュ等）。
- Chores
  - 依存更新（bcryptjs）と検索パフォーマンス計測スクリプト、.gitignore更新を追加。
<!-- end of auto-generated comment: release notes by coderabbit.ai -->